pyproject.toml
[tool.poetry]
name = "mydjangoportfolio"
version = "0.1.0"
description = "My portfolio made in Django. Made for practice and showcase of my skills."
authors = ["Adam Boudník <boudnikadam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.1"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
django-modeltranslation = "^0.18.11"
pillow = "^10.2.0"
django-markdownfield = "^0.11.0"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
whitenoise = "^6.6.0"
django-storages = {extras = ["s3"], version = "^1.14.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
